criando o primeiro mapa e correção de bugs

+++ src/entity/Entity.java
@@ -3,7 +3,7 @@ package entity;
 public class Entity {
 
     public int x, y;
-    public int aderencia;
+    public double aderencia;
     public double aceleracao;
     public double velocidade=0;
     public int limVelocidade;
